20_08_20_before/requirement.txt
alabaster==0.7.12
alembic==1.2.1
aniso8601==8.0.0
app==0.0.1
argh==0.26.2
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bleach==3.1.0
bootstrap4==0.1.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==4.0.2
cryptography==2.8
cycler==0.10.0
dataclasses==0.6
DateTime==4.3
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.6
django-bootstrap==0.2.4
django-bootstrap4==1.0.1
django-crispy-forms==1.8.1
django-extensions==2.2.5
docopt==0.6.2
docutils==0.15.2
dominate==2.4.0
entrypoints==0.3
et-xmlfile==1.0.1
Faker==2.0.3
flake8==3.7.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
flask-paginate==0.5.4
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
future==0.18.2
get==2019.4.13
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipyparallel==6.2.4
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mpld3==0.3
mysqlclient==1.4.4
nba-api==1.1.4
nbconvert==5.6.1
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numpy==1.17.3
openpyxl==3.0.1
packaging==19.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pathtools==0.1.2
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pip-date==1.0.3
plotly==4.3.0
pluggy==0.13.0
popong-nlp==0.1
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
public==2019.4.13
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylava==0.2.2
pylint==2.4.3
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==5.2.2
pytest-watch==4.2.0
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-editor==1.0.4
python-jsonrpc-server==0.2.0
python-language-server==0.30.0
pytrends==4.7.2
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
query-string==2019.4.13
regex==2019.11.1
request==2019.4.13
requests==2.22.0
retrying==1.3.3
rope==0.14.0
rsa==4.0
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
soupsieve==1.9.4
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
stevedore==1.31.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
uWSGI==2.0.18
viridis==0.4.2
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
visitor==0.1.3
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
yapf==0.28.0
zipp==0.6.0
zope.interface==4.6.0
